+++ packages/modules/package.js
@@ -6,11 +6,10 @@ Package.describe({
 });
 
 Npm.depends({
-  reify: "0.11.21"
+  reify: "0.11.22"
 });
 
 Package.onUse(function(api) {
-  api.use("underscore");
   api.use("modules-runtime");
   api.mainModule("client.js", "client");
   api.mainModule("server.js", "server");